anyworkx_backend/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
channels==4.0.0
channels-redis==4.0.0
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
msgpack==1.0.5
Pillow==9.4.0
PyJWT==2.6.0
pytz==2022.7.1
redis==4.5.1
sqlparse==0.4.3
